+++ src/canister/src/logic/report_logic.rs
@@ -1,6 +1,5 @@
 use canister_types::models::{
     api_error::ApiError,
-    filter_type::FilterType,
     paged_response::PagedResponse,
     report::{PostReport, Report, ReportFilter, ReportResponse, ReportSort},
 };
@@ -8,6 +7,8 @@ use ic_cdk::caller;
 use std::collections::HashMap;
 
 use crate::storage::{MemberStore, ReportStore, StorageMethods};
+
+use super::profile_logic::ProfileCalls;
 
 pub struct ReportCalls;
 
@@ -20,59 +21,51 @@ impl ReportCalls {
         }
 
         let report = Report::new(post_report);
-        let result = ReportStore::insert(report);
-        ReportResponse::from_result(result)
+        let result = ReportStore::insert(report)?;
+        ReportResponse::from_result(
+            Ok(result.clone()),
+            ProfileCalls::get_subject_response_by_subject(&result.1.subject),
+        )
     }
 
     pub fn get_report(report_id: u64) -> Result<ReportResponse, ApiError> {
-        ReportResponse::from_result(ReportStore::get(report_id))
+        let result = ReportStore::get(report_id)?;
+        ReportResponse::from_result(
+            Ok(result.clone()),
+            ProfileCalls::get_subject_response_by_subject(&result.1.subject),
+        )
     }
 
     pub fn get_reports(
         limit: usize,
         page: usize,
         sort: ReportSort,
-        filters: Vec<FilterType<ReportFilter>>,
+        filters: Vec<ReportFilter>,
         group_id: u64,
     ) -> Result<PagedResponse<ReportResponse>, ApiError> {
-        let reports =
-            ReportStore::filter(|_, report| report.group_id.is_some_and(|id| id == group_id));
+        let mut reports =
+            ReportStore::filter(|_, report| report.group_id.is_some_and(|id| id == group_id))
+                .into_iter()
+                .collect::<HashMap<u64, Report>>();
 
-        // split the filters into or and and filters
-        let mut or_filters: Vec<ReportFilter> = vec![];
-        let mut and_filters: Vec<ReportFilter> = vec![];
-        for filter_type in filters {
-            use FilterType::*;
-            match filter_type {
-                And(filter_value) => and_filters.push(filter_value),
-                Or(filter_value) => or_filters.push(filter_value),
-            }
-        }
-
-        // filter the groups based on the `OR` filters
-        let mut or_filtered_reports: HashMap<u64, Report> = HashMap::new();
-        for filter in or_filters {
-            for (id, report) in &reports {
-                if filter.is_match(id, report) {
-                    or_filtered_reports.insert(id.clone(), report.clone());
+        for filter in filters {
+            for (id, report) in &reports.clone() {
+                if !filter.is_match(id, report) {
+                    reports.remove(&id);
                 }
             }
         }
 
-        // filter the `or_filtered` groups based on the `AND` filters
-        let mut and_filtered_reports: HashMap<u64, Report> = HashMap::new();
-        for filter in and_filters {
-            for (id, report) in &or_filtered_reports {
-                if filter.is_match(id, report) {
-                    and_filtered_reports.insert(id.clone(), report.clone());
-                }
-            }
-        }
-
-        let sorted_reports = sort.sort(and_filtered_reports);
+        let sorted_reports = sort.sort(reports);
         let result: Vec<ReportResponse> = sorted_reports
             .into_iter()
-            .map(|data| ReportResponse::new(data.0, data.1))
+            .map(|data| {
+                ReportResponse::new(
+                    data.0,
+                    data.1.clone(),
+                    ProfileCalls::get_subject_response_by_subject(&data.1.subject),
+                )
+            })
             .collect();
 
         Ok(PagedResponse::new(page, limit, result))