+++ src/canister/src/storage/friend_request_storage.rs
@@ -1,5 +1,6 @@
-use super::storage_api::{StorageMethods, FRIEND_REQUEST};
+use super::storage_api::{StorageMethods, CLEAR_MEMORY_ID, FRIEND_REQUEST, MEMORY_MANAGER};
 use canister_types::models::{api_error::ApiError, friend_request::FriendRequest};
+use ic_stable_structures::StableBTreeMap;
 
 pub struct FriendRequestStore;
 
@@ -134,5 +135,14 @@ impl StorageMethods<u64, FriendRequest> for FriendRequestStore {
     /// # Note
     fn remove(key: u64) -> bool {
         FRIEND_REQUEST.with(|data| data.borrow_mut().remove(&key).is_some())
+    }
+
+    /// Clear all attendees
+    fn clear() -> () {
+        FRIEND_REQUEST.with(|n| {
+            n.replace(StableBTreeMap::new(
+                MEMORY_MANAGER.with(|m| m.borrow().get(CLEAR_MEMORY_ID)),
+            ))
+        });
     }
 }