+++ src/canister/src/storage/storage_api.rs
@@ -13,6 +13,8 @@ use std::cell::RefCell;
 
 pub type Memory = VirtualMemory<DefaultMemoryImpl>;
 
+pub static CLEAR_MEMORY_ID: MemoryId = MemoryId::new(254);
+
 /// The memory IDs for the different stores.
 /// # Note
 /// These IDs are used to identify the different stores in the `MemoryManager`.
@@ -72,6 +74,7 @@ pub trait StorageMethods<K, V> {
     fn insert_by_key(key: K, entity: V) -> Result<(K, V), ApiError>;
     fn update(id: K, entity: V) -> Result<(K, V), ApiError>;
     fn remove(id: K) -> bool;
+    fn clear() -> ();
 }
 
 /// A trait for the identifier reference maps.